Add files via upload
Intelligent packaging is a relatively recent innovation that aligns with sustainable development. By employing active and interactive technologies it manages to monitor food quality, improve food safety, and record product information across the supply chain. By incorporating intelligent tools, such as barcodes, time-temperature indicators, and gas sensors, intelligent packaging transforms conventional packaging conspicuously. 

Previous studies have tried to investigate how customers perceive and understand intelligent packaging. However, not many studies have examined the factors that affect the adoption of intelligent packaging. This study fills the gap and examines the factors that affect the adoption of intelligent packaging by deploying the Diffusion of Innovation theory. Based on the Diffusion of Innovation theory and the data collected from 106 supermarket customers, it is found that relative advantage, compatibility, observability, and trialability have a positive effect on the adoption of intelligent packaging. On the contrary, complexity and perceived value pricing have a negative effect on the adoption of this innovation.